chore: move error state
