+++ apps/web/src/components/Home/Suggested.tsx
@@ -1,6 +1,6 @@
 import Loader from '@components/Shared/Loader';
 import UserProfile from '@components/Shared/UserProfile';
-import { ErrorMessage } from '@components/UI/ErrorMessage';
+import { ErrorMessage } from 'ui';
 import { UsersIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import type { Profile } from 'lens';
