+++ apps/web/src/components/Shared/Modal/Collectors.tsx
@@ -1,6 +1,6 @@
 import UserProfile from '@components/Shared/UserProfile';
 import WalletProfile from '@components/Shared/WalletProfile';
-import { ErrorMessage } from '@components/UI/ErrorMessage';
+import { ErrorMessage } from 'ui';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import type { Profile, Wallet, WhoCollectedPublicationRequest } from 'lens';
