+++ apps/web/src/components/Shared/Modal/Likes.tsx
@@ -1,5 +1,5 @@
 import UserProfile from '@components/Shared/UserProfile';
-import { ErrorMessage } from '@components/UI/ErrorMessage';
+import { ErrorMessage } from 'ui';
 import { HeartIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import type { Profile, WhoReactedPublicationRequest } from 'lens';
