+++ apps/web/src/components/Shared/Modal/Mirrors.tsx
@@ -1,5 +1,5 @@
 import UserProfile from '@components/Shared/UserProfile';
-import { ErrorMessage } from '@components/UI/ErrorMessage';
+import { ErrorMessage } from 'ui';
 import { SwitchHorizontalIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import type { Profile, ProfileQueryRequest } from 'lens';
