+++ packages/ui/index.ts
@@ -3,4 +3,5 @@ export { Button } from './Button';
 export { Card } from './Card';
 export { Checkbox } from './Checkbox';
 export { EmptyState } from './EmptyState';
+export { ErrorMessage } from './ErrorMessage';
 export { FieldError, Form, useZodForm } from './Form';
